+++ client/commands/compile.js
@@ -34,9 +34,6 @@ module.exports = function compile({ debug = false, useStartUp = false } = {}) {
 
   if (process.status === 0) {
     outputFile = replaceFileExtension(fileToCompile, ".prg");
-    if (debug) {
-      createBreakpointsFile(outputDir, outputFile);
-    }
   } else {
     vscode.window.showErrorMessage("Compilation failed with errors.");
     output.append(process.stderr.toString());
@@ -56,26 +53,4 @@ function findStartUp(file) {
   if (startUp) {
     return path.join(fileDir, startUp);
   }
-}
-
-function createBreakpointsFile(outputDir, outputFile) {
-  const viceSymbols = path.join(outputDir, replaceFileExtension(outputFile, ".vs"));
-  if (!fs.existsSync(viceSymbols)) {
-    return;
-  }
-
-  const breakpoints = fs
-    .readFileSync(viceSymbols, { encoding: "utf8" })
-    .split("\n")
-    .filter((x) => x.startsWith("break"));
-
-  if (breakpoints.length === 0) {
-    output.appendLine("No breakpoints found, skipping.");
-    return;
-  }
-
-  const breakpointsFile = path.join(outputDir, replaceFileExtension(outputFile, ".breakpoints"));
-
-  fs.writeFileSync(breakpointsFile, `${breakpoints.join("\n")}\n`);
-  output.appendLine(`Wrote breakpoints to ${breakpointsFile}`);
 }